Document the magnificent SymbolIndexer.
