+++ metaserver/src/main/scala/scala/meta/languageserver/Workspace.scala
@@ -29,5 +29,4 @@ object Workspace {
       }
     )
   }
-
 }
